Replaced std::list with std::vector and fixed a critical bug.
With the std::vector the program seems to run better, still need to test elm::vector with its fast elm::vector::erase().

Fixed a bug related to corrupted pointers. It happened when the container needed to reallocate the elements, all the pointers got invalidated. Fixed that and made BlockPool::findFirstFreeBlock() faster by retrieving the block's index in the vector instead of returning a reference to the block.